requirements.txt
streamlit==1.28.1
tensorflow==2.13.0
keras==2.13.1
numpy==1.24.3
pillow==10.0.1
opencv-python==4.8.1.78
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
tensorboard==2.13.0